Address PR feedback, rename file to tables_constants.hpp
