+++ sdk/tables/azure-data-tables/src/private/tables_constants.hpp
