+++ sdk/tables/azure-data-tables/src/table_clients.cpp
@@ -9,7 +9,7 @@
 #include "private/policies/tenant_bearer_token_policy.hpp"
 #include "private/policies/timeout_policy.hpp"
 #include "private/serializers.hpp"
-#include "private/table_constants.hpp"
+#include "private/tables_constants.hpp"
 
 #include <sstream>
 #include <string>
